Candidato Finish, Ciudadano Finish, Votacion Finish, Informes Init

+++ src/main/java/com/ingSoft/InfoVotantes/candidato/service/serviceImpl/CandidatoServiceImpl.java
@@ -2,59 +2,82 @@ package com.ingSoft.InfoVotantes.candidato.service.serviceImpl;
 
 import com.ingSoft.InfoVotantes.candidato.component.CandidatoConverter;
 import com.ingSoft.InfoVotantes.candidato.dto.CandidatoDTO;
+import com.ingSoft.InfoVotantes.candidato.model.Candidato;
 import com.ingSoft.InfoVotantes.candidato.repository.CandidatoRepository;
 import com.ingSoft.InfoVotantes.candidato.service.CandidatoService;
 import com.ingSoft.InfoVotantes.ciudadano.model.Ciudadano;
 import com.ingSoft.InfoVotantes.ciudadano.repository.CiudadanoRepository;
-
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.var;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import lombok.var;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class CandidatoServiceImpl implements CandidatoService {
-    @Autowired
-    private CandidatoRepository candidatoRepository;
-    @Autowired
-    private CandidatoConverter candidatoConverter;
-    @Autowired
-    private CiudadanoRepository ciudadanoRepository;
+    private final CandidatoRepository candidatoRepository;
+    private final CandidatoConverter candidatoConverter;
+    private final CiudadanoRepository ciudadanoRepository;
+
+    public CandidatoServiceImpl(CandidatoRepository candidatoRepository, CandidatoConverter candidatoConverter, CiudadanoRepository ciudadanoRepository) {
+        this.candidatoRepository = candidatoRepository;
+        this.candidatoConverter = candidatoConverter;
+        this.ciudadanoRepository = ciudadanoRepository;
+    }
 
     @Override
     public Object add(CandidatoDTO candidatoDto) {
-        Optional<Ciudadano> personaCiudadano = ciudadanoRepository.findById(Integer.parseInt(candidatoDto.getCed_ciudadano()));
+        Optional<Ciudadano> personaCiudadano = ciudadanoRepository.findById(candidatoDto.getCedCiudadano());
         if(personaCiudadano.isPresent()){
-            var candidato = candidatoConverter.dtoToModel(candidatoDto);
-            candidato.setCed_ciudadano(personaCiudadano.get());
+            Optional<Candidato> personaCandidato = candidatoRepository.findByCedCiudadano(candidatoDto.getCedCiudadano());
+            if(!personaCandidato.isPresent()){
+                var candidato = candidatoConverter.dtoToModel(candidatoDto);
+                candidato.setCedCiudadano(personaCiudadano.get());
+                return candidatoRepository.save(candidato);
+            }
+            return new ResponseEntity<>("The Candidato is registered", HttpStatus.UNAUTHORIZED);
+        }
+        return new ResponseEntity<>("The Ciudadano isn't registered", HttpStatus.UNAUTHORIZED);
+    }
+
+    @Override
+    public List<Candidato> getAll() {
+        return candidatoRepository.findAll();
+    }
+
+    @Override
+    public Object readCandidato(String id) {
+        Optional<Candidato> personaCandidato = candidatoRepository.findById(id);
+        if(personaCandidato.isPresent()){
+            return personaCandidato.get();
+        }
+        return new ResponseEntity<>("The Candidato isn't registered", HttpStatus.UNAUTHORIZED);
+    }
+
+    @Override
+    public Object update(String id, CandidatoDTO candidatoDto) {
+        Optional<Candidato> personaCandidato = candidatoRepository.findById(id);
+
+        if(personaCandidato.isPresent()){
+            Candidato candidato = personaCandidato.get();
+            candidato.setNombre(candidatoDto.getNombre());
+            candidato.setEslogan(candidatoDto.getEslogan());
+            candidato.setPropuesta(candidatoDto.getPropuesta());
+            candidato.setPartidoPolitico(candidatoDto.getPartidoPolitico());
             return candidatoRepository.save(candidato);
         }
-        return new ResponseEntity<>("The User isn't registered", HttpStatus.UNAUTHORIZED);
+        return new ResponseEntity<>("Cant find any Candidato with the given id", HttpStatus.UNAUTHORIZED);
     }
 
     @Override
-    public List<CandidatoDTO> getAll() {
-        var candidatos = candidatoRepository.findAll();
-        return candidatoConverter.listModelToDto(candidatos);
-    }
-
-    @Override
-    public CandidatoDTO readCandidato(int id) {
-        var cliente = candidatoRepository.getCandidatoById(id);
-        return candidatoConverter.modelToDto(cliente);
-    }
-
-    @Override
-    public Object update(String cedula, CandidatoDTO candidatoDto) {
-        return null;
-    }
-
-    @Override
-    public void delete(String id) {
-        candidatoRepository.deleteById(id);
+    public Object delete(String id) {
+        Optional<Candidato> personaCandidato = candidatoRepository.findById(id);
+        if(personaCandidato.isPresent()){
+            candidatoRepository.deleteById(id);
+            return new ResponseEntity<>("User Delete", HttpStatus.UNAUTHORIZED);
+        }
+        return new ResponseEntity<>("Cant find any Candidato with the given id", HttpStatus.UNAUTHORIZED);
     }
 }
